Beim ersten Mal starten der Applikation, gab es eine NullPointerException, da eine Notification abgebrochen worden ist, die noch nicht vorhanden war.

+++ app/src/main/java/at/htl/medassistant/MainActivity.java
@@ -222,15 +222,21 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
+        boolean someTreatmentsInList = false;
+        if (treatmentList.size() > 0) {
+            someTreatmentsInList = true;
+        }
         treatmentList.clear();
         if (db.getTreatments() != null) {
             treatmentList.addAll(db.getTreatments());
             adapter.notifyDataSetChanged();
-
         }
 
         try {
-            boundService.destroyAndCreateNewNotification();
+            if (someTreatmentsInList) {
+                boundService.destroyAndCreateNewNotification();
+            }
+
         } catch (Exception e) {
             e.printStackTrace();
         }
